# openapi.yaml
openapi: 3.0.3
info:
  title: Zomaal Backend API
  description: |
    REST API for the Zomaal backend (MVP). Successful JSON responses use `{ "success": true, "data": ... }`.
    Errors use `{ "success": false, "error": { "code": "...", "message": "..." } }` (in development, `stack` may be included on 500s).

    **Store isolation:** Authenticated business routes use the JWT `storeId`; do not trust a client-supplied store id in the body for authorization.

    **Order pricing:** Line unit prices are computed on the server from product price tiers; the client must not send per-line prices.

    **Protected routes** (except auth + health): `Authorization: Bearer <accessToken>`. Store-scoped routes also rely on `storeScope` + `storeMiddleware` (JWT `storeId` must match any explicit `storeId` in path/header/query/body).

    Placeholder routes: `/users`, `/integrations`, parts of `/stores`, `/transactions`, legacy `/kpi/*`.

    Implemented: **orders** (customer object + warnings), **returns** (scan, manual, status), **customers** (risk + blacklisted), **dashboard/kpis**, **expenses**, **alerts**.
  version: 1.1.0
servers:
  - url: http://localhost:5000/api
    description: Local (default PORT from env, else 5000)
  - url: https://{host}/api
    description: Deployed
    variables:
      host:
        default: api.example.com
...
paths:
...
  /products:
    post:
      tags: [Products]
      summary: Create product with unit-price tiers
      description: |
        Each price tier: `quantity` = minimum units for that tier; `price` = unit price when that tier applies.
        For a line quantity, the tier with the **largest** `quantity` such that `tier.quantity <= line quantity` wins.
      operationId: createProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "409":
          description: SKU conflict in store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    get:
      tags: [Products]
      summary: List products (paginated)
      operationId: listProducts
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: includeArchived
          in: query
          description: Pass `true` or `1` to include archived products.
          schema:
            type: string
            enum: ["true", "false", "1", "0"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListSuccessResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
  /products/{id}:
    parameters:
      - $ref: "#/components/parameters/ResourceId"
    patch:
      tags: [Products]
      summary: Partial update product
      description: Optional full replacement of `prices` when `prices` is sent (must be non-empty).
      operationId: updateProduct
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Products]
      summary: Archive product (soft delete)
      description: Sets `isArchived` and deactivates; does not hard-delete.
      operationId: archiveProduct
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSuccessResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /customers:
    post:
      tags: [Customers]
      summary: Create or find customer by phone/email
      description: At least one of `phone` or `email` is required. Returns `201` if created, `200` if existing, with `meta.wasCreated`.
      operationId: upsertCustomer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertCustomerRequest"
      responses:
        "200":
          description: Existing customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerUpsertSuccessResponse"
        "201":
          description: New customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerUpsertSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Customers]
      summary: List customers
      operationId: listCustomers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListSuccessResponse"
        "401":
          $ref: "#/components/responses/ErrorResponse"
...
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Get customer by id
      operationId: getCustomer
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSuccessResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
  /returns/scan:
    post:
      tags: [Returns]
      summary: Scan return (match order or create NEEDS_REVIEW)
      description: |
        Provide at least one of `trackingId` (matches `Order.externalId`), `orderReference` (UUID → order id, else externalId), or `phone` (latest order for customer).
        Matched: `verificationStatus` VERIFIED, `scannedAt` set, return line items copied from order. Unmatched: `orderId` null, `NEEDS_REVIEW`, `scanMetadata`.
      operationId: scanReturn
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnScanRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
  /returns/manual:
    post:
      tags: [Returns]
      summary: Manual return (quantities must not exceed order lines)
      operationId: manualReturn
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReturnManualRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
  /returns/{id}/status:
    patch:
      tags: [Returns]
      summary: Update return status (REFUSED increments customer refusalCount)
      operationId: updateReturnStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  $ref: "#/components/schemas/ReturnStatus"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
        "404":
          $ref: "#/components/responses/ErrorResponse"
...
  /expenses:
    post:
      tags: [Expenses]
      summary: Create expense
      operationId: createExpense
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExpenseRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseSuccessResponse"
        "400":
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Expenses]
      summary: List expenses (paginated)
      operationId: listExpenses
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseListResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Access JWT payload (minimal): `userId`, `role` (store role for this token), `storeId`.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    ErrorResponse:
      description: Client or server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            stack:
              type: string
              description: Present in development for 500 errors
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
...
    OrderStatus:
      type: string
      description: Order lifecycle status; PATCH must follow allowed transitions from current status.
      enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED, RETURNED]
    ReturnStatus:
      type: string
      enum: [PENDING, RECEIVED, REFUSED, CANCELLED]
    VerificationStatus:
      type: string
      enum: [VERIFIED, UNVERIFIED, NEEDS_REVIEW]
    ExpenseType:
      type: string
      enum: [ADS, PURCHASE, SHIPPING, OTHER]
    ProductPriceTier:
      type: object
      required: [quantity, price]
      properties:
        quantity:
          type: integer
          minimum: 1
          description: Minimum units for this tier
        price:
          type: number
          format: float
          exclusiveMinimum: 0
          description: Unit price when tier applies
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        title:
          type: string
        sku:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        isActive:
          type: boolean
        isArchived:
          type: boolean
        stock:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
        prices:
          type: array
          items:
            $ref: "#/components/schemas/ProductPriceRow"
    ProductPriceRow:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        quantity:
          type: integer
        price:
          type: number
          format: float
    CreateProductRequest:
      type: object
      required: [title, prices]
      properties:
        title:
          type: string
        sku:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        stock:
          type: integer
          minimum: 0
          default: 0
        lowStockThreshold:
          type: integer
          minimum: 0
          default: 5
        prices:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProductPriceTier"
    UpdateProductRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
        sku:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        stock:
          type: integer
          minimum: 0
        lowStockThreshold:
          type: integer
          minimum: 0
        isActive:
          type: boolean
        prices:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ProductPriceTier"
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        isBlacklisted:
          type: boolean
        riskLevel:
          type: integer
        totalOrdersCount:
          type: integer
        returnCount:
          type: integer
        cancelCount:
          type: integer
        refusalCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    UpsertCustomerRequest:
      type: object
      description: At least one of phone or email required.
      properties:
        phone:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
...
    ReturnScanRequest:
      type: object
      description: At least one field required.
      properties:
        trackingId:
          type: string
        orderReference:
          type: string
        phone:
          type: string
    ReturnManualRequest:
      type: object
      required: [orderId, items]
      properties:
        orderId:
          type: string
          format: uuid
        items:
          type: array
          minItems: 1
          items:
            type: object
            required: [productId, quantity]
            properties:
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
                minimum: 1
        status:
          $ref: "#/components/schemas/ReturnStatus"
        reason:
          type: string
        comment:
          type: string
    Return:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
          nullable: true
        status:
          $ref: "#/components/schemas/ReturnStatus"
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        reason:
          type: string
          nullable: true
        comment:
          type: string
          nullable: true
        scannedAt:
          type: string
          format: date-time
          nullable: true
        scanMetadata:
          nullable: true
          description: Present when scan could not match an order (opaque JSON).
        createdAt:
          type: string
          format: date-time
        order:
          nullable: true
          type: object
          properties:
            id:
              type: string
              format: uuid
            status:
              $ref: "#/components/schemas/OrderStatus"
            externalId:
              type: string
              nullable: true
            totalAmount:
              type: number
              format: float
            customerId:
              type: string
              format: uuid
              nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              returnId:
                type: string
                format: uuid
              productId:
                type: string
                format: uuid
              quantity:
                type: integer
              product:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  sku:
                    type: string
                    nullable: true
...
    Expense:
      type: object
      properties:
        id:
          type: string
          format: uuid
        storeId:
          type: string
          format: uuid
        type:
          $ref: "#/components/schemas/ExpenseType"
        amount:
          type: number
          format: float
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateExpenseRequest:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          description: Use lowercase in JSON — `ads`, `purchase`, `shipping`, `other` (mapped to enum server-side).
          enum: [ads, purchase, shipping, other]
        amount:
          type: number
          format: float
          exclusiveMinimum: 0
        note:
          type: string
...
    ProductSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Product"
    ProductListSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Product"
            pagination:
              $ref: "#/components/schemas/Pagination"
    CustomerSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Customer"
    CustomerUpsertSuccessResponse:
      type: object
      required: [success, data, meta]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Customer"
        meta:
          type: object
          required: [wasCreated]
          properties:
            wasCreated:
              type: boolean
    CustomerListSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Customer"
            pagination:
              $ref: "#/components/schemas/Pagination"
...
    ReturnSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Return"
...
    ExpenseSuccessResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          $ref: "#/components/schemas/Expense"
    ExpenseListResponse:
      type: object
      required: [success, data]
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Expense"
            pagination:
              $ref: "#/components/schemas/Pagination"
...